Merge pull request #535 from Difegue/dev

+++ lib/LANraragi/Utils/Archive.pm
@@ -15,8 +15,8 @@ use Redis;
 use Cwd;
 use Data::Dumper;
 use Image::Magick;
-use Archive::Peek::Libarchive;
-use Archive::Extract::Libarchive;
+use Archive::Libarchive::Extract;
+use Archive::Libarchive::Peek;
 
 use LANraragi::Utils::TempFolder qw(get_temp);
 use LANraragi::Utils::Logging qw(get_logger);
@@ -25,7 +25,8 @@ use LANraragi::Utils::Generic qw(is_image shasum);
 # Utilitary functions for handling Archives.
 # Relies on Libarchive, ImageMagick and GhostScript for PDFs.
 use Exporter 'import';
-our @EXPORT_OK = qw(is_file_in_archive extract_file_from_archive extract_archive extract_thumbnail generate_thumbnail);
+our @EXPORT_OK =
+  qw(is_file_in_archive extract_file_from_archive extract_single_file extract_archive extract_thumbnail generate_thumbnail get_filelist);
 
 sub is_pdf {
     my ( $filename, $dirs, $suffix ) = fileparse( $_[0], qr/\.[^.]*/ );
@@ -46,58 +47,86 @@ sub generate_thumbnail {
     undef $img;
 }
 
-#extract_archive(path, archive_to_extract)
-#Extract the given archive to the given path.
+# sanitize_filename(filename)
+# Converts extracted filenames to an ascii variant to avoid extra filesystem headaches.
+sub sanitize_filename {
+
+    my $filename = $_[0];
+    eval {
+        # Try a guess to regular japanese encodings first
+        $filename = decode( "Guess", $filename );
+    };
+
+    # Fallback to utf8
+    $filename = decode_utf8($filename) if $@;
+
+    # Re-encode the result to ASCII and move the file to said result name.
+    # Use Encode's coderef feature to map non-ascii characters to their Unicode codepoint equivalent.
+    $filename = encode( "ascii", $filename, sub { sprintf "%04X", shift } );
+
+    if ( length $filename > 254 ) {
+        $filename = substr( $filename, 0, 254 );
+    }
+
+    return $filename;
+}
+
+# extract_archive(path, archive_to_extract, force)
+# Extract the given archive to the given path.
+# This sub won't re-extract files already present in the destination unless force = 1.
 sub extract_archive {
 
-    my ( $destination, $to_extract ) = @_;
+    my ( $destination, $to_extract, $force_extract ) = @_;
+    my $logger = get_logger( "Archive", "lanraragi" );
 
     # PDFs are handled by Ghostscript (alas)
     if ( is_pdf($to_extract) ) {
         return extract_pdf( $destination, $to_extract );
     }
 
-    # build an Archive::Extract object
-    my $ae = Archive::Extract::Libarchive->new( archive => $to_extract );
+    # Prepare libarchive with a callback to skip over existing files (unless force=1)
+    my $ae = Archive::Libarchive::Extract->new(
+        filename => $to_extract,
+        entry    => sub {
+            my $e = shift;
+            if ($force_extract) { return 1; }
 
-    #Extract to $destination. Report if it fails.
-    my $ok = $ae->extract( to => $destination ) or die $ae->error;
+            my $filename = $e->pathname;
+            $filename = sanitize_filename($filename);
+            if ( -e "$destination/$filename" ) {
+                $logger->debug("$filename already exists in $destination");
+                return 0;
+            }
 
-    my $cwd = getcwd();
+            # Pre-emptively create the file to signal we're working on it
+            open( my $fh, ">", "$destination/$filename" ) or return 0;
+            close $fh;
+            return 1;
+        }
+    );
+
+    # Extract to $destination. This method throws if extraction fails.
+    $ae->extract( to => $destination );
+
+    # Get extraction folder
+    my $result_dir = $ae->to;
+    my $cwd        = getcwd();
 
     # Rename extracted files and folders to an encoded version for easier handling
     finddepth(
         sub {
             unless ( $_ eq '.' ) {
-
-                my $filename = $_;
-                eval {
-                    # Try a guess to regular japanese encodings first
-                    $filename = decode( "Guess", $filename );
-                };
-
-                # Fallback to utf8
-                $filename = decode_utf8($filename) if $@;
-
-                # Re-encode the result to ASCII and move the file to said result name.
-                # Use Encode's coderef feature to map non-ascii characters to their Unicode codepoint equivalent.
-                $filename = encode( "ascii", $filename, sub { sprintf "%04X", shift } );
-
-                if ( length $filename > 254 ) {
-                    $filename = substr( $filename, 0, 254 );
-                }
-
-                move( $_, $filename );
+                move( $_, sanitize_filename($_) );
             }
         },
-        $ae->extract_path
+        $result_dir
     );
 
     # chdir back to the base cwd in case finddepth died midway
     chdir $cwd;
 
-    # dir that was extracted to
-    return $ae->extract_path;
+    # Return the directory we extracted the files to.
+    return $result_dir;
 }
 
 sub extract_pdf {
@@ -126,11 +155,12 @@ sub extract_pdf {
     return $destination;
 }
 
-#extract_thumbnail(thumbnaildir, id)
-#Finds the first image for the specified archive ID and makes it the thumbnail.
+# extract_thumbnail(thumbnaildir, id)
+# Finds the first image for the specified archive ID and makes it the thumbnail.
 sub extract_thumbnail {
 
     my ( $thumbdir, $id ) = @_;
+    my $logger = get_logger( "Archive", "lanraragi" );
 
     # Another subfolder with the first two characters of the id is used for FS optimization.
     my $subfolder = substr( $id, 0, 2 );
@@ -140,28 +170,31 @@ sub extract_thumbnail {
     my $redis = LANraragi::Model::Config->get_redis;
 
     my $file = $redis->hget( $id, "file" );
-    my $temppath = get_temp . "/thumb/$id/";
+    my $temppath = get_temp . "/thumb/$id";
 
     # Make sure the thumb temp dir exists
     make_path($temppath);
 
-    my $arcimg = "";
-    if ( is_pdf($file) ) {
-        $arcimg = extract_page_pdf( $file, $temppath );
-    } else {
-        $arcimg = extract_page_libarchive( $file, $temppath );
-    }
+    # Get first image from archive using filelist
+    my @filelist    = get_filelist($file);
+    my $first_image = $filelist[0];
 
-    #While we have the image, grab its SHA-1 hash for tag research.
-    #That way, no need to repeat the costly extraction later.
+    die "First image not found" unless $first_image;
+    $logger->debug("Extracting thumbnail for $id from $first_image");
+
+    # Extract first image to temp dir
+    my $arcimg = extract_single_file( $file, $first_image, $temppath );
+
+    # While we have the image, grab its SHA-1 hash for tag research.
+    # That way, no need to repeat the costly extraction later.
     my $shasum = shasum( $arcimg, 1 );
     $redis->hset( $id, "thumbhash", $shasum );
     $redis->quit();
 
-    #Thumbnail generation
+    # Thumbnail generation
     generate_thumbnail( $arcimg, $thumbname );
 
-    #Delete the previously extracted file.
+    # Delete the previously extracted file.
     unlink $arcimg;
 
     # Clean up safe folder
@@ -169,65 +202,40 @@ sub extract_thumbnail {
     return $thumbname;
 }
 
-sub extract_page_pdf {
+# get_filelist($archive)
+# Returns a list of all the files contained in the given archive.
+sub get_filelist {
 
-    my ( $file, $temppath ) = @_;
-    mkdir $temppath;
-    my $output = $temppath . "pdf_first_page.jpg";
+    my $archive = $_[0];
+    my @files   = ();
 
-    my $logger = get_logger( "Archive", "lanraragi" );
+    if ( is_pdf($archive) ) {
 
-    my $gscmd = "gs -dNOPAUSE -dLastPage=1 -sDEVICE=jpeg -r72 -o '$output' '$file'";
-    $logger->debug("Sending PDF $file to GhostScript...");
-    $logger->debug($gscmd);
+        # For pdfs, extraction returns images from 1.jpg to x.jpg, where x is the pdf pagecount.
+        my $pages = `gs -q -c "($archive) (r) file runpdfbegin pdfpagecount = quit"`;
+        for my $num ( 1 .. $pages ) {
+            push @files, "$num.jpg";
+        }
+    } else {
+        my $peek = Archive::Libarchive::Peek->new( filename => $archive );
 
-    `$gscmd`;
-
-    return $output;
-}
-
-sub extract_page_libarchive {
-
-    my ( $file, $temppath ) = @_;
-
-    # Get all the files of the archive
-    my $peek = Archive::Peek::Libarchive->new( filename => $file );
-    my @files = $peek->files();
-    my @extracted;
-
-    # Filter out non-images
-    foreach my $file (@files) {
-        if ( is_image($file) ) {
-            push @extracted, $file;
+        # Filter out non-images
+        foreach my $file ( $peek->files ) {
+            if ( is_image($file) ) {
+                push @files, $file;
+            }
         }
     }
 
-    @extracted = sort { lc($a) cmp lc($b) } @extracted;
-
-    # Get the first file of the list and spit it out into a file
-    my $contents = $peek->file( $extracted[0] );
-
-    # The name sometimes comes with the folder as a bonus, so we use basename to filter it out.
-    my ( $filename, $dirs, $suffix ) = fileparse( $extracted[0], qr/\.[^.]*/ );
-
-    # Move the extracted file to a safe folder to avoid concurrent overwrites
-    my $arcimg = $temppath . $filename . $suffix;
-
-    open( my $fh, '>', $arcimg )
-      or die "Could not open file '$arcimg' $!";
-    print $fh $contents;
-    close $fh;
-
-    return $arcimg;
+    return @files;
 }
 
-#is_file_in_archive($archive, $file)
-#Uses libarchive::peek to figure out if $archive contains $file.
-#Returns 1 if it does exist, 0 otherwise.
+# is_file_in_archive($archive, $file)
+# Uses libarchive::peek to figure out if $archive contains $file.
+# Returns 1 if it does exist, 0 otherwise.
 sub is_file_in_archive {
 
     my ( $archive, $wantedname ) = @_;
-
     my $logger = get_logger( "Archive", "lanraragi" );
 
     if ( is_pdf($archive) ) {
@@ -238,7 +246,7 @@ sub is_file_in_archive {
     $logger->debug("Iterating files of archive $archive, looking for '$wantedname'");
     $Data::Dumper::Useqq = 1;
 
-    my $peek = Archive::Peek::Libarchive->new( filename => $archive );
+    my $peek = Archive::Libarchive::Peek->new( filename => $archive );
     my $found = 0;
     $peek->iterate(
         sub {
@@ -254,42 +262,64 @@ sub is_file_in_archive {
     return $found;
 }
 
-#extract_file_from_archive($archive, $file)
-#Extract $file from $archive and returns the filesystem path it's extracted to.
-#If the file doesn't exist in the archive, this will still create a file, but empty.
+# extract_single_file ($archive, $file, $destination)
+# Extract $file from $archive to $destination and returns the filesystem path it's extracted to.
+# If the file doesn't exist in the archive, this will still create a file, but empty.
+sub extract_single_file {
+
+    my ( $archive, $filepath, $destination ) = @_;
+    my $logger = get_logger( "Archive", "lanraragi" );
+
+    my $outfile = "$destination/$filepath";
+    $logger->debug("Output for single file extraction: $outfile");
+
+    # Remove file from $outfile and hand the full directory to make_path
+    my ( $name, $path, $suffix ) = fileparse( $outfile, qr/\.[^.]*/ );
+    make_path($path);
+
+    if ( is_pdf($archive) ) {
+
+        # For pdfs the filenames are always x.jpg, so we pull the page number from that
+        my $page = $filepath;
+        $page =~ s/^(\d+).jpg$/$1/;
+
+        my $gscmd = "gs -dNOPAUSE -dFirstPage=$page -dLastPage=$page -sDEVICE=jpeg -r200 -o '$outfile' '$archive'";
+        $logger->debug("Extracting page $filepath from PDF $archive");
+        $logger->debug($gscmd);
+
+        `$gscmd`;
+    } else {
+
+        my $contents = "";
+        my $peek = Archive::Libarchive::Peek->new( filename => $archive );
+        $contents = $peek->file($filepath);
+
+        open( my $fh, '>', $outfile )
+          or die "Could not open file '$outfile' $!";
+        print $fh $contents;
+        close $fh;
+    }
+
+    my $fixed_name = sanitize_filename($outfile);
+    move( $outfile, $fixed_name );
+    return $fixed_name;
+}
+
+# extract_file_from_archive($archive, $file)
+# Variant for plugins.
+# Extracts the file with a timestamp to a folder in /temp/plugin.
 sub extract_file_from_archive {
 
     my ( $archive, $filename ) = @_;
 
-    #Timestamp extractions in microseconds
+    # Timestamp extractions in microseconds
     my ( $seconds, $microseconds ) = gettimeofday;
     my $stamp = "$seconds-$microseconds";
     my $path  = get_temp . "/plugin/$stamp";
     mkdir get_temp . "/plugin";
     mkdir $path;
-    my $contents = "";
 
-    unless ( is_pdf($archive) ) {
-        my $peek = Archive::Peek::Libarchive->new( filename => $archive );
-        $peek->iterate(
-            sub {
-                my ( $file, $data ) = @_;
-
-                if ( $file =~ /$filename$/ ) {
-                    $contents = $data;
-                }
-            }
-        );
-    }
-
-    my $outfile = $path . "/" . $filename;
-
-    open( my $fh, '>', $outfile )
-      or die "Could not open file '$outfile' $!";
-    print $fh $contents;
-    close $fh;
-
-    return $outfile;
+    return extract_single_file( $archive, $filename, $path );
 }
 
 1;
